Add os.listdir() for note tracking
Changed the install.py file to add '.notes' and '.archive_notes'
directories inside of the .quicknote directory. Eliminated '.filelist'
and '.archivelist' and instead now loop through directories using
listdir() to return the names of the included files (which are notes).
Changed every method that relied on .filelist (which was most of the
note.py methods) to be independent of that. Now need to keep track
of fewer files and don't have to constantly be reading and writing
to filelist. Also added some more error-checking when appropriate
while going through to remove deprecated calls to .filelist.