Add archive_note and un_archive_note
Added the ability to archive notes (transfers them to a separate
directory called '.archive_notes' inside of the .quicknote directory)
as well as the ability to reverse this and unarchive notes. This functionality
will allow people to archive notes for later use, without having to see these
notes in their main feed.