Add renameNote and clearArchiveNotes
Added implementations of renameNote and clearArchiveNotes. Fixed
the issue where getCurrentNote couldn't read multi word filenames.
However, getCurrentNote adds unreadable characters to the path
when reading in dataFile so that needs to be fixed still (many
functions are not working bc they getCurrentNote is not working).
Also, removed background_info and made sure that functions
take fewer parameters in general and do more of the heavy lifting
on their own to reduce overhead speed slowdowns.