Remove deprecated python files
Removed the python_implementation directory as well as the qnote
file and qnote.rb file. The whole project has been rewritten in C
so it felt unnecessary to hold on to deprecated code. If the time
comes where it seems necessary to have this code, can look previously
in git history.