Remove ruby file and add first version of merge
Removed the Formula directory because liszt-music/homebrew-liszt is
the new repo where homebrew users can tap this repository. Changed
the name of the current repo to just 'liszt'. Adding a first version
of note merging in mergeNotes(). mergeNotes() is a bit buggy, and
doesn't change to default if the current note is merged. Also,
updated the readme to contain the new brew install info as well
as how to download lst through a git clone.