Add JSON functionality
Finally was able to get cJSON files to be included properly so as
to avoid linker errors. For whatever reason the cJSON.c file hadn't
been included in the cJSON.h file by its create Dave Gamble so I fixed
that. Added JSON functionality throughout the project so that whenever
information is written to or taken from the data file it is done through
JSON. This makes things immensely simpler. Going to merge this branch now.